requirements.txt
boto3
cython
numpy
Pillow
git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI